\name{AHLRBaseDbs-package}
\alias{AHLRBaseDbs-package}
\alias{AHLRBaseDbs}
\docType{package}
\title{
\packageTitle{AHLRBaseDbs}
}
\description{
\packageDescription{AHLRBaseDbs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AHLRBaseDbs}
\packageIndices{AHLRBaseDbs}
}
\author{
\packageAuthor{AHLRBaseDbs}

Maintainer: \packageMaintainer{AHLRBaseDbs}
}

\keyword{ package }
\examples{
ls("package:AHLRBaseDbs") # This is empty package
}